requirements.txt
Flask==3.0.0
Pillow==10.2.0
gunicorn==21.2.0 